14. In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,15 +7,15 @@
 int main(void)
 {
 	int i;
-	int t[50];
+	long int t[50];
 
 	t[0] = 1;
 	t[1] = 2;
-	printf("%d, %d", t[0], t[1]);
+	printf("%ld, %ld", t[0], t[1]);
 	for (i = 0; i < 48; i++)
 	{
 		t[i + 2] = t[i] + t[i + 1];
-		printf(", %d ", t[i + 2]);
+		printf(", %ld ", t[i + 2]);
 	}
 	printf("\n");
 	return (0);
